+++ authentication-server/server.ts
@@ -99,11 +99,16 @@ async function boot (){
         app.post("/user/login", validationMiddleware.login, authController.userLogin);
         app.post("/org/login", validationMiddleware.login, authController.organizationLogin);
 
+        app.post("/user/forgotPassword", validationMiddleware.forgotPassword, userController.forgotPassword);
+        app.post("/org/forgotPassword", validationMiddleware.forgotPassword, orgController.forgotPassword);
+
         app.get("/sessions", authMiddleware.checkAuthToken, authController.getSessions);
         app.delete("/revokeSession", authMiddleware.checkAuthToken, validationMiddleware.validateRevokeSession, authController.revokeSession);
         app.delete("/revokeAllSessions", authMiddleware.checkAuthToken, authController.revokeAllSessions);
 
         app.put("/verifyEmail/:magicToken", validationMiddleware.verifyEmail, authController.verifyEmail);
+
+        app.patch("/resetPassword/:magicToken", validationMiddleware.resetPassword, authController.resetPassword);
 
         app.get("/logout", authMiddleware.checkAuthToken, authController.logout);