+++ frontend-server/src/App.tsx
@@ -15,8 +15,11 @@ const UserLoginPage = lazy(() => authPages().then(module => ({ default: module.U
 const OrganizationLoginPage = lazy(() => authPages().then(module => ({ default: module.OrganizationLoginPage })));
 const UserSignupPage = lazy(() => authPages().then(module => ({ default: module.UserSignupPage })));
 const OrganizationSignupPage = lazy(() => authPages().then(module => ({ default: module.OrganizationSignupPage })));
+const UserForgotPasswordPage = lazy(() => authPages().then(module => ({ default: module.UserForgotPasswordPage })));
+const OrganizationForgotPasswordPage = lazy(() => authPages().then(module => ({ default: module.OrganizationForgotPasswordPage })));
 const ConsentPage = lazy(() => authPages().then(module => ({ default: module.ConsentPage })));
-const EmailVerificationPage = lazy(() => authPages().then(module => ({ default: module.verifyEmailPage })));
+const EmailVerificationPage = lazy(() => authPages().then(module => ({ default: module.VerifyEmailPage })));
+const ResetPasswordPage = lazy(() => authPages().then(module => ({ default: module.ResetPasswordPage })));
 const ClientsPage = lazy(() => import("@/features/organization/OrganizationPages"));
 const NotFoundPage = lazy(() => import("@/features/shared/SharedPages").then(module => ({ default: module.NotFoundPage })));
 
@@ -31,7 +34,10 @@ function AppRoutes() {
 					<Route path={APP_ROUTES.organization.login} element={<OrganizationLoginPage />} />
 					<Route path={APP_ROUTES.user.signup} element={<UserSignupPage />} />
 					<Route path={APP_ROUTES.organization.signup} element={<OrganizationSignupPage />} />
+					<Route path={APP_ROUTES.user.forgotPassword} element={<UserForgotPasswordPage />} />
+					<Route path={APP_ROUTES.organization.forgotPassword} element={<OrganizationForgotPasswordPage />} />
 					<Route path={APP_ROUTES.emailVerification} element={<EmailVerificationPage/>} />
+					<Route path={APP_ROUTES.resetPassword} element={<ResetPasswordPage />} />
 				</Route>
 
 				<Route element={<ProtectedRoute entityType="user" />}>