+++ frontend-server/src/config/routes.ts
@@ -4,15 +4,18 @@ export const APP_ROUTES = {
 	root: "/",
 	consent: "/consent",
 	emailVerification: "/verify-email/:magicToken",
+	resetPassword: "/reset-password/:magicToken",
 	user: {
 		login: "/user/login",
 		signup: "/user/signup",
+		forgotPassword: "/user/forgot-password",
 		portal: "/user/portal",
 		session: "/user/portal/session",
 	},
 	organization: {
 		login: "/org/login",
 		signup: "/org/signup",
+		forgotPassword: "/org/forgot-password",
 		portal: "/org/portal",
 		clients: "/org/portal/clients",
 		session: "/org/portal/session",
@@ -24,6 +27,7 @@ export const API_ROUTES = {
 	logout: "/logout",
 	consent: "/consent",
 	verifyEmail: "/verifyEmail",
+	resetPassword: "/resetPassword",
 	authorize: "/authorize",
 	session: "/sessions",
 	revokeSession:"/revokeSession",
@@ -32,11 +36,13 @@ export const API_ROUTES = {
 		login: "/user/login",
 		signup: "/user/signup",
 		profile: "/user/profile",
+		forgotPassword: "/user/forgotPassword"
 	},
 	organization: {
 		login: "/org/login",
 		signup: "/org/signup",
 		clients: "/org/clients",
+		forgotPassword: "/org/forgotPassword"
 	},
 };