+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The Following rules must be followed at all times
 - Added initial release documentation with proper structure ( `CHANGELOG.md` and `RELEASE_NOTES.md`)
 - Added Email Verification feature
 - Added Session management feature
+- Added Forgot and Reset Password feature
 
 ### Changed
 - Improvised the UI for all Pages